+++ projects/gateway2/translator/plugins/httplisteneroptions/query/query.go
@@ -2,14 +2,10 @@ package query
 
 import (
 	"context"
-	"errors"
-	"fmt"
 
 	solokubev1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1/kube/apis/gateway.solo.io/v1"
 	"github.com/solo-io/gloo/projects/gateway2/translator/plugins/utils"
-	"github.com/solo-io/gloo/projects/gateway2/wellknown"
 	skv2corev1 "github.com/solo-io/skv2/pkg/api/core.skv2.solo.io/v1"
-	"k8s.io/apimachinery/pkg/fields"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
@@ -59,74 +55,40 @@ func NewQuery(c client.Client) HttpListenerOptionQueries {
 	return &queries{c}
 }
 
-func (q *queries) GetAttachedHttpListenerOptions(
+func (r *queries) GetAttachedHttpListenerOptions(
 	ctx context.Context,
 	listener *gwv1.Listener,
 	parentGw *gwv1.Gateway,
 	parentListenerSet *apixv1a1.XListenerSet,
 ) ([]*solokubev1.HttpListenerOption, error) {
-	if parentGw == nil {
-		return nil, errors.New("nil parent gateway")
-	}
-	if parentGw.GetName() == "" || parentGw.GetNamespace() == "" {
-		return nil, fmt.Errorf("parent gateway must have name and namespace; received name: %s, namespace: %s", parentGw.GetName(), parentGw.GetNamespace())
+	createList := func() *solokubev1.HttpListenerOptionList {
+		return &solokubev1.HttpListenerOptionList{}
 	}
 
-	nnk := utils.NamespacedNameKind{
-		Namespace: parentGw.Namespace,
-		Name:      parentGw.Name,
-		Kind:      wellknown.GatewayKind,
-	}
-
-	listGw := &solokubev1.HttpListenerOptionList{}
-	if err := q.c.List(
-		ctx,
-		listGw,
-		client.MatchingFieldsSelector{Selector: fields.OneTermEqualSelector(HttpListenerOptionTargetField, nnk.String())},
-		client.InNamespace(parentGw.GetNamespace()),
-	); err != nil {
-		return nil, err
-	}
-
-	listListenerSet := &solokubev1.HttpListenerOptionList{}
-	if parentListenerSet != nil {
-		nnkListenerSet := utils.NamespacedNameKind{
-			Namespace: parentListenerSet.GetNamespace(),
-			Name:      parentListenerSet.GetName(),
-			Kind:      wellknown.XListenerSetKind,
+	// Can't just do this in the function because we need to call `list.Items`
+	extractItems := func(list *solokubev1.HttpListenerOptionList) []*solokubev1.HttpListenerOption {
+		items := make([]*solokubev1.HttpListenerOption, len(list.Items))
+		for i, item := range list.Items {
+			items[i] = &item
 		}
-
-		if err := q.c.List(
-			ctx,
-			listListenerSet,
-			client.MatchingFieldsSelector{Selector: fields.OneTermEqualSelector(HttpListenerOptionTargetField, nnkListenerSet.String())},
-			client.InNamespace(parentListenerSet.GetNamespace()),
-		); err != nil {
-			return nil, err
-		}
+		return items
 	}
 
-	allItems := append(listGw.Items, listListenerSet.Items...)
-	if len(allItems) == 0 {
-		return nil, nil
-	}
-
-	policies := buildWrapperType(allItems)
-	orderedPolicies := utils.GetPrioritizedListenerPolicies(policies, listener, parentGw.Name, parentListenerSet)
-	return orderedPolicies, nil
-}
-
-func buildWrapperType(
-	list []solokubev1.HttpListenerOption,
-) []utils.PolicyWithSectionedTargetRefs[*solokubev1.HttpListenerOption] {
-	policies := []utils.PolicyWithSectionedTargetRefs[*solokubev1.HttpListenerOption]{}
-	for i := range list {
-		item := &list[i]
-
-		policy := httpListenerOptionPolicy{
+	wrapPolicy := func(item *solokubev1.HttpListenerOption) utils.PolicyWithSectionedTargetRefs[*solokubev1.HttpListenerOption] {
+		return httpListenerOptionPolicy{
 			obj: item,
 		}
-		policies = append(policies, policy)
 	}
-	return policies
+
+	return utils.GetOptionsForListener(
+		context.Background(),
+		listener,
+		parentGw,
+		parentListenerSet,
+		r.c,
+		HttpListenerOptionTargetField,
+		createList,
+		extractItems,
+		wrapPolicy,
+	)
 }
