+++ projects/gateway2/translator/plugins/listeneroptions/query/query.go
@@ -2,13 +2,10 @@ package query
 
 import (
 	"context"
-	"fmt"
 
 	solokubev1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1/kube/apis/gateway.solo.io/v1"
 	"github.com/solo-io/gloo/projects/gateway2/translator/plugins/utils"
-	"github.com/solo-io/gloo/projects/gateway2/wellknown"
 	skv2corev1 "github.com/solo-io/skv2/pkg/api/core.skv2.solo.io/v1"
-	"k8s.io/apimachinery/pkg/fields"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
@@ -61,64 +58,34 @@ func (r *listenerOptionQueries) GetAttachedListenerOptions(
 	parentGw *gwv1.Gateway,
 	parentListenerSet *apixv1a1.XListenerSet,
 ) ([]*solokubev1.ListenerOption, error) {
-	if parentGw.GetName() == "" || parentGw.GetNamespace() == "" {
-		return nil, fmt.Errorf("parent gateway must have name and namespace; received name: %s, namespace: %s", parentGw.GetName(), parentGw.GetNamespace())
+	createList := func() *solokubev1.ListenerOptionList {
+		return &solokubev1.ListenerOptionList{}
 	}
 
-	nnk := utils.NamespacedNameKind{
-		Namespace: parentGw.Namespace,
-		Name:      parentGw.Name,
-		Kind:      wellknown.GatewayKind,
-	}
-
-	list := &solokubev1.ListenerOptionList{}
-	if err := r.c.List(
-		ctx,
-		list,
-		client.MatchingFieldsSelector{Selector: fields.OneTermEqualSelector(ListenerOptionTargetField, nnk.String())},
-		client.InNamespace(parentGw.GetNamespace()),
-	); err != nil {
-		return nil, err
-	}
-
-	listListenerSet := &solokubev1.ListenerOptionList{}
-	if parentListenerSet != nil {
-		nnkListenerSet := utils.NamespacedNameKind{
-			Namespace: parentListenerSet.GetNamespace(),
-			Name:      parentListenerSet.GetName(),
-			Kind:      wellknown.XListenerSetKind,
+	// Can't just do this in the function because we need to call `list.Items`
+	extractItems := func(list *solokubev1.ListenerOptionList) []*solokubev1.ListenerOption {
+		items := make([]*solokubev1.ListenerOption, len(list.Items))
+		for i, item := range list.Items {
+			items[i] = &item
 		}
-		if err := r.c.List(
-			ctx,
-			listListenerSet,
-			client.MatchingFieldsSelector{Selector: fields.OneTermEqualSelector(ListenerOptionTargetField, nnkListenerSet.String())},
-			client.InNamespace(parentListenerSet.GetNamespace()),
-		); err != nil {
-			return nil, err
-		}
+		return items
 	}
 
-	allItems := append(list.Items, listListenerSet.Items...)
-	if len(allItems) == 0 {
-		return nil, nil
-	}
-
-	policies := buildWrapperType(allItems)
-	orderedPolicies := utils.GetPrioritizedListenerPolicies(policies, listener, parentGw.Name, parentListenerSet)
-	return orderedPolicies, nil
-}
-
-func buildWrapperType(
-	items []solokubev1.ListenerOption,
-) []utils.PolicyWithSectionedTargetRefs[*solokubev1.ListenerOption] {
-	policies := []utils.PolicyWithSectionedTargetRefs[*solokubev1.ListenerOption]{}
-	for i := range items {
-		item := &items[i]
-
-		policy := listenerOptionPolicy{
+	wrapPolicy := func(item *solokubev1.ListenerOption) utils.PolicyWithSectionedTargetRefs[*solokubev1.ListenerOption] {
+		return listenerOptionPolicy{
 			obj: item,
 		}
-		policies = append(policies, policy)
 	}
-	return policies
+
+	return utils.GetOptionsForListener(
+		context.Background(),
+		listener,
+		parentGw,
+		parentListenerSet,
+		r.c,
+		ListenerOptionTargetField,
+		createList,
+		extractItems,
+		wrapPolicy,
+	)
 }
