+++ projects/gateway2/translator/plugins/virtualhostoptions/query/query.go
@@ -3,13 +3,9 @@ package query
 import (
 	"context"
 
-	"github.com/rotisserie/eris"
 	solokubev1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1/kube/apis/gateway.solo.io/v1"
 	"github.com/solo-io/gloo/projects/gateway2/translator/plugins/utils"
-	"github.com/solo-io/gloo/projects/gateway2/wellknown"
 	skv2corev1 "github.com/solo-io/skv2/pkg/api/core.skv2.solo.io/v1"
-	"k8s.io/apimachinery/pkg/fields"
-	"k8s.io/apimachinery/pkg/types"
 	apixv1a1 "sigs.k8s.io/gateway-api/apisx/v1alpha1"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -55,83 +51,6 @@ func NewQuery(c client.Client) VirtualHostOptionQueries {
 	return &virtualHostOptionQueries{c}
 }
 
-func (r *virtualHostOptionQueries) GetVirtualHostOptionsForListener0(
-	ctx context.Context,
-	listener *gwv1.Listener,
-	parentGw *gwv1.Gateway,
-	parentListenerSet *apixv1a1.XListenerSet,
-) ([]*solokubev1.VirtualHostOption, error) {
-	if parentGw.GetName() == "" || parentGw.GetNamespace() == "" {
-		return nil, eris.Errorf("parent gateway must have name and namespace; received name: %s, namespace: %s", parentGw.GetName(), parentGw.GetNamespace())
-	}
-
-	nnk := utils.NamespacedNameKind{
-		Namespace: parentGw.Namespace,
-		Name:      parentGw.Name,
-		Kind:      wellknown.GatewayKind,
-	}
-
-	listGw := &solokubev1.VirtualHostOptionList{}
-	if err := r.c.List(
-		ctx,
-		listGw,
-		client.MatchingFieldsSelector{Selector: fields.AndSelectors(
-			fields.OneTermEqualSelector(VirtualHostOptionTargetField, nnk.String()),
-		)},
-		client.InNamespace(parentGw.GetNamespace()),
-	); err != nil {
-		return nil, err
-	}
-
-	listListenerSet := &solokubev1.VirtualHostOptionList{}
-	if parentListenerSet != nil {
-		nnkListenerSet := utils.NamespacedNameKind{
-			Namespace: parentListenerSet.GetNamespace(),
-			Name:      parentListenerSet.GetName(),
-			Kind:      wellknown.XListenerSetKind,
-		}
-
-		if err := r.c.List(
-			ctx,
-			listListenerSet,
-			client.MatchingFieldsSelector{Selector: fields.AndSelectors(
-				fields.OneTermEqualSelector(VirtualHostOptionTargetField, nnkListenerSet.String()),
-			)},
-			client.InNamespace(parentListenerSet.GetNamespace()),
-		); err != nil {
-			return nil, err
-		}
-	}
-
-	allItems := append(listGw.Items, listListenerSet.Items...)
-	if len(allItems) == 0 {
-		return nil, nil
-	}
-
-	policies := buildWrapperType(allItems)
-	orderedPolicies := utils.GetPrioritizedListenerPolicies(policies, listener, parentGw.Name, parentListenerSet)
-	return orderedPolicies, nil
-}
-
-func buildWrapperType(
-	items []solokubev1.VirtualHostOption,
-) []utils.PolicyWithSectionedTargetRefs[*solokubev1.VirtualHostOption] {
-	policies := []utils.PolicyWithSectionedTargetRefs[*solokubev1.VirtualHostOption]{}
-	for i := range items {
-		item := &items[i]
-
-		policy := vhostOptionPolicy{
-			obj: item,
-		}
-		policies = append(policies, policy)
-	}
-	return policies
-}
-
-type OptionsList interface {
-	client.ObjectList
-}
-
 func (r *virtualHostOptionQueries) GetVirtualHostOptionsForListener(
 	ctx context.Context,
 	listener *gwv1.Listener,
@@ -157,7 +76,7 @@ func (r *virtualHostOptionQueries) GetVirtualHostOptionsForListener(
 		}
 	}
 
-	return GetOptionsForListener(
+	return utils.GetOptionsForListener(
 		context.Background(),
 		listener,
 		parentGw,
@@ -168,83 +87,4 @@ func (r *virtualHostOptionQueries) GetVirtualHostOptionsForListener(
 		extractItems,
 		wrapPolicy,
 	)
-}
-
-// Use to eliminate `extractItems` from `GetOptionsForListener`
-type ObjectListWithItems[T client.Object] interface {
-	client.ObjectList
-	GetItems() []T
-}
-
-func GetOptionsForListener[T client.Object, TList client.ObjectList](
-	ctx context.Context,
-	listener *gwv1.Listener,
-	parentGw *gwv1.Gateway,
-	parentListenerSet *apixv1a1.XListenerSet,
-	c client.Client,
-	optionTargetField string,
-	createList func() TList,
-	extractItems func(list TList) []T,
-	wrapPolicy func(item T) utils.PolicyWithSectionedTargetRefs[T],
-) ([]T, error) {
-	if parentGw.GetName() == "" || parentGw.GetNamespace() == "" {
-		return nil, eris.Errorf("parent gateway must have name and namespace; received name: %s, namespace: %s", parentGw.GetName(), parentGw.GetNamespace())
-	}
-
-	parentListenerSetName := ""
-	if parentListenerSet != nil {
-		parentListenerSetName = parentListenerSet.GetName()
-	}
-
-	nn := types.NamespacedName{
-		Namespace: parentGw.Namespace,
-		Name:      parentGw.Name,
-	}
-
-	nnListenerSet := types.NamespacedName{
-		Namespace: parentGw.Namespace,
-		Name:      parentListenerSetName,
-	}
-
-	listGw := createList()
-	if err := c.List(
-		ctx,
-		listGw,
-		client.MatchingFieldsSelector{Selector: fields.OneTermEqualSelector(optionTargetField, nn.String())},
-		client.InNamespace(parentGw.GetNamespace()),
-	); err != nil {
-		return nil, err
-	}
-
-	listListenerSet := createList()
-	if parentListenerSet != nil {
-		if err := c.List(
-			ctx,
-			listListenerSet,
-			client.MatchingFieldsSelector{Selector: fields.OneTermEqualSelector(optionTargetField, nnListenerSet.String())},
-		); err != nil {
-			return nil, err
-		}
-	}
-
-	allItems := []T{}
-	allItems = append(allItems, extractItems(listGw)...)
-	allItems = append(allItems, extractItems(listListenerSet)...)
-
-	policies := buildWrapperGeneric(allItems, wrapPolicy)
-	orderedPolicies := utils.GetPrioritizedListenerPolicies(policies, listener, parentGw.Name, parentListenerSet)
-	return orderedPolicies, nil
-}
-
-func buildWrapperGeneric[T client.Object](
-	items []T,
-	wrapPolicy func(item T) utils.PolicyWithSectionedTargetRefs[T],
-) []utils.PolicyWithSectionedTargetRefs[T] {
-	policies := []utils.PolicyWithSectionedTargetRefs[T]{}
-	for i := range items {
-		item := items[i]
-		policy := wrapPolicy(item)
-		policies = append(policies, policy)
-	}
-	return policies
 }
